third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/editing/style/bold-with-dom-changes-expected.txt

CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
This test verifies that deleting DOM elements in a DOMSubtreeModified handler triggered by and execCommand('bold') does not produce a crash.
PASS
